Re-structured files into new folders for better readability
